backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
drf-yasg[validation]>=1.21,<2.0
psycopg2-binary>=2.9,<3.0
python-jose[cryptography]>=3.3,<4.0
boto3>=1.34,<2.0
requests>=2.31,<3.0
redis>=5.0,<6.0
django-environ>=0.11,<1.0
gunicorn>=21.2,<22.0
whitenoise>=6.6,<7.0
